requirements.txt
asn1crypto==1.5.1
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.4
cryptography==40.0.2
filelock==3.12.2
idna==3.4
mypy-extensions==1.0.0
oscrypto==1.3.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.1
pycparser==2.21
pycryptodomex==3.18.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
snowflake-connector-python==3.0.4
sortedcontainers==2.4.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==1.26.16
